I am finish a cart and create a bottom sheet.

+++ lib/data/models/favorite_m.dart
@@ -7,7 +7,7 @@ class FavoriteModel {
   String? itemsImage;
   int? itemsCount;
   int? itemsActive;
-  int? itemsPrice;
+  double? itemsPrice;
   int? itemsDiscount;
   String? itemsDate;
   int? itemsCategories;
